refactor: the Setup class to registerign blocks
